requirements.txt
imposm.parser==1.0.7
python-graph-core==1.8.2
wheel==0.26.0
